Couple of fixes after testing
