+++ src/utilities/index.js
@@ -1,6 +1,4 @@
 import * as arraysUtils from './arrays'
-import * as animationUtils from './animation'
-import * as browserUtils from './browser'
 import * as classNamesUtils from './classNames'
 import * as collectionUtils from './collection'
 import * as colorUtils from './color'
@@ -17,8 +15,6 @@ import * as stringsUtils from './strings'
 import * as urlUtils from './urls'
 
 export const arrays = arraysUtils
-export const animation = animationUtils
-export const browser = browserUtils
 export const classNames = classNamesUtils
 export const color = colorUtils
 export const collection = collectionUtils
@@ -36,8 +32,6 @@ export const urls = urlUtils
 
 export default {
   arrays,
-  animation,
-  browser,
   classNames,
   collection,
   color,
